split tree view out
