+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Materials/MaterialEditor.cs
@@ -1,12 +1,8 @@
-using System;
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEditor.IMGUI.Controls;
 using UnityEngine;
-using VisualPinball.Engine.Game;
-using VisualPinball.Engine.VPT;
 using VisualPinball.Unity.Extensions;
-using VisualPinball.Unity.VPT;
 using VisualPinball.Unity.VPT.Table;
 
 namespace VisualPinball.Unity.Editor.Materials
@@ -16,7 +12,7 @@ namespace VisualPinball.Unity.Editor.Materials
 	/// </summary>
 	public class MaterialEditor : EditorWindow
 	{
-		private TreeViewTest _treeView;
+		private MaterialListView _listView;
 		private TreeViewState _treeViewState;
 
 		private bool _foldoutVisual = true;
@@ -62,11 +58,18 @@ namespace VisualPinball.Unity.Editor.Materials
 			}
 
 			EditorGUILayout.BeginHorizontal();
+			GUILayout.Button("Add", GUILayout.ExpandWidth(false));
+			GUILayout.Button("Remove", GUILayout.ExpandWidth(false));
+			GUILayout.Button("Clone", GUILayout.ExpandWidth(false));
+			EditorGUILayout.EndHorizontal();
+
+			EditorGUILayout.BeginHorizontal();
 
 			// list
 			GUILayout.FlexibleSpace();
 			var r = GUILayoutUtility.GetLastRect();
-			_treeView.OnGUI(new Rect(0, 0, r.width, position.height));
+			var listRect = new Rect(r.x, r.y, r.width, position.height - r.y);
+			_listView.OnGUI(listRect);
 
 			// options
 			EditorGUILayout.BeginVertical(GUILayout.MaxWidth(300));
@@ -78,7 +81,7 @@ namespace VisualPinball.Unity.Editor.Materials
 						Undo.RecordObject(_table, "Rename Material");
 						_selectedMaterial.Name = _renameBuffer;
 						_renaming = false;
-						_treeView.Reload();
+						_listView.Reload();
 					}
 					if (GUILayout.Button("Cancel")) {
 						_renaming = false;
@@ -195,15 +198,15 @@ namespace VisualPinball.Unity.Editor.Materials
 
 		private void UndoPerformed()
 		{
-			if (_treeView != null) {
-				_treeView.Reload();
+			if (_listView != null) {
+				_listView.Reload();
 			}
 		}
 
 		private void FindTable()
 		{
 			_table = GameObject.FindObjectOfType<TableBehavior>();
-			_treeView = new TreeViewTest(_treeViewState, _table, MaterialSelected);
+			_listView = new MaterialListView(_treeViewState, _table, MaterialSelected);
 		}
 
 		private void MaterialSelected(List<Engine.VPT.Material> selectedMaterials)
@@ -214,184 +217,6 @@ namespace VisualPinball.Unity.Editor.Materials
 				_renaming = false;
 			}
 			Repaint();
-		}
-	}
-
-	class TreeViewTest : TreeView // TODO: rename and move to its own file
-	{
-		public event Action<List<Engine.VPT.Material>> MaterialSelected;
-
-		private TableBehavior _table;
-
-		public TreeViewTest(TreeViewState treeViewState, TableBehavior table, Action<List<Engine.VPT.Material>> materialSelected) : base(treeViewState)
-		{
-			MaterialSelected += materialSelected;
-			_table = table;
-
-			var columns = new[]
-			{
-				new MultiColumnHeaderState.Column
-				{
-					headerContent = new GUIContent("Name"),
-					headerTextAlignment = UnityEngine.TextAlignment.Left,
-					canSort = true,
-					sortedAscending = true,
-					sortingArrowAlignment = UnityEngine.TextAlignment.Right,
-					width = 300,
-					minWidth = 100,
-					maxWidth = float.MaxValue,
-					autoResize = true,
-					allowToggleVisibility = false,
-				},
-				new MultiColumnHeaderState.Column
-				{
-					headerContent = new GUIContent("In use"),
-					headerTextAlignment = UnityEngine.TextAlignment.Left,
-					canSort = true,
-					sortedAscending = true,
-					sortingArrowAlignment = UnityEngine.TextAlignment.Right,
-					width = 50,
-					minWidth = 50,
-					maxWidth = 50,
-					autoResize = true,
-					allowToggleVisibility = false,
-				},
-			};
-
-			var headerState = new MultiColumnHeaderState(columns);
-			this.multiColumnHeader = new MultiColumnHeader(headerState);
-			this.multiColumnHeader.SetSorting(0, true);
-			this.multiColumnHeader.sortingChanged += SortingChanged;
-			this.showAlternatingRowBackgrounds = true;
-			this.showBorder = true;
-
-			Reload();
-			if (GetRows().Count > 0) {
-				SetSelection(new List<int> { 0 }, TreeViewSelectionOptions.FireSelectionChanged);
-			}
-		}
-
-		private void SortingChanged(MultiColumnHeader multiColumnHeader)
-		{
-			Reload();
-		}
-
-		public override void OnGUI(Rect rect)
-		{
-			// if the table went away, force a rebuild to empty out the list
-			if (_table == null && GetRows().Count > 0) {
-				Reload();
-			}
-			base.OnGUI(rect);
-		}
-
-		protected override TreeViewItem BuildRoot()
-		{
-			return new TreeViewItem { id = -1, depth = -1, displayName = "Root" };
-		}
-
-		protected override IList<TreeViewItem> BuildRows(TreeViewItem root)
-		{
-			var items = new List<TreeViewItem>();
-			if (_table == null) return items;
-
-			// collect list of in use materials
-			List<string> inUseMaterials = new List<string>();
-			var renderables = _table.GetComponentsInChildren<IItemBehaviorWithMaterials>();
-			foreach (var renderable in renderables) {
-				var mats = renderable.UsedMaterials;
-				if (mats != null) {
-					foreach (var mat in mats) {
-						if (!string.IsNullOrEmpty(mat)) {
-							inUseMaterials.Add(mat);
-						}
-					}
-				}
-			}
-
-			// get row data for each material
-			for (int i = 0; i < _table.Item.Data.Materials.Length; i++) {
-				var mat = _table.Item.Data.Materials[i];
-				items.Add(new RowData(i, mat, inUseMaterials.Contains(mat.Name)));
-			}
-
-			var sortedColumns = this.multiColumnHeader.state.sortedColumns;
-			if (sortedColumns.Length > 0) {
-				items.Sort((baseA, baseB) => {
-					var a = baseA as RowData;
-					var b = baseB as RowData;
-					// sort based on multiple columns
-					foreach (var column in sortedColumns) {
-						bool ascending = multiColumnHeader.IsSortedAscending(column);
-						// flip for descending
-						if (!ascending) {
-							var tmp = b;
-							b = a;
-							a = tmp;
-						}
-						int compareResult = 0;
-						switch (column) {
-							case 0:
-								compareResult = a.Material.Name.CompareTo(b.Material.Name);
-								break;
-							case 1:
-								compareResult = a.InUse.CompareTo(b.InUse);
-								break;
-						}
-						// not equal in this column, then return that
-						if (compareResult != 0) {
-							return compareResult;
-						}
-					}
-					return a.CompareTo(b);
-				});
-			}
-
-			return items;
-		}
-
-		protected override void RowGUI(RowGUIArgs args)
-		{
-			for (int i = 0; i < args.GetNumVisibleColumns(); ++i) {
-				CellGUI(args.GetCellRect(i), args.item, args.GetColumn(i));
-			}
-		}
-
-		private void CellGUI(Rect cellRect, TreeViewItem item, int column)
-		{
-			CenterRectUsingSingleLineHeight(ref cellRect);
-			var rowData = item as RowData;
-			switch (column) {
-				case 0: // todo: make an enum for the columns
-					GUI.Label(cellRect, rowData.Material.Name);
-					break;
-				case 1:
-					GUI.Label(cellRect, rowData.InUse ? "X" : "");
-					break;
-			}
-		}
-
-		protected override void SelectionChanged(IList<int> selectedIds)
-		{
-			List<Engine.VPT.Material> selectedMats = new List<Engine.VPT.Material>();
-			var rows = GetRows();
-			foreach (var row in rows) {
-				if (selectedIds.Contains(row.id)) {
-					selectedMats.Add((row as RowData).Material);
-				}
-			}
-			MaterialSelected?.Invoke(selectedMats);
-		}
-
-		private class RowData : TreeViewItem
-		{
-			public readonly Engine.VPT.Material Material;
-			public readonly bool InUse;
-
-			public RowData(int id, Engine.VPT.Material mat, bool inUse) : base(id, 0) {
-				Material = mat;
-				InUse = inUse;
-			}
 		}
 	}
 }
