+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Materials/MaterialListView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60313a75f2a01dd4c87ef096b3d505f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
